rag-agent/pyproject.toml
[tool.poetry]
name = "chat-with-your-data-solution-accelerator"
version = "1.1.0"
description = "Chat with your data solution accelerator"
authors = []
readme = "README.md"
packages = [
    { include = "code" }
]

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "1.24.0"
streamlit = "1.54.0"
python-dotenv = "1.2.1"
azure-ai-formrecognizer = "3.3.3"
azure-storage-blob = "12.28.0"
azure-identity = "1.25.2"
flask = {extras = ["async"], version = "^3.1.2"}
openai = "1.109.1"
langchain = "0.2.17"
langchain-community = "0.2.19"
langchain-openai = "0.1.25"
requests = "2.32.5"
tiktoken = "0.12.0"
azure-storage-queue = "12.15.0"
beautifulsoup4 = "4.14.3"
fake-useragent = "2.2.0"
chardet = "5.2.0"
azure-search-documents = "11.6.0b1"
azure-ai-contentsafety = "1.0.0"
python-docx = "1.2.0"
azure-keyvault-secrets = "4.10.0"
pandas = "2.3.3"
azure-monitor-opentelemetry = "^1.6.10"
opentelemetry-instrumentation-httpx = "^0.52b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.30.0"
pillow = "11.0.0"
azure-mgmt-cognitiveservices = "^14.1.0"
jsonschema = "^4.25.1"
semantic-kernel = {version = "1.39.0", python = "<3.13"}
pydantic = "2.11.10"
azure-ai-ml = "^1.31.0"
azure-cosmos = "^4.14.3"
asyncpg = "^0.31.0"
psycopg2-binary = "^2.9.11"
pgvector = "^0.4.2"
rfc3987 = "^1.3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "7.0.0"
flake8 = "7.3.0"
black = "25.12.0"
pre-commit = "4.5.1"
pytest_httpserver = "1.1.5"
trustme = "1.2.1"
jupyter = "1.1.1"
pytest-asyncio = "^1.2.0"
azure-ai-evaluation = ">=0.5.0"

[tool.poetry.group.prompt-flow]
optional = true

[tool.poetry.group.prompt-flow.dependencies]
promptflow = {extras = ["azure"], version = "1.18.1"}

[tool.coverage.run]
omit = [
    "**/tests/*",
    "**/dist/*",
    "**/infra/*",
]

[tool.coverage.report]
skip_empty = true
include_namespace_packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
